backend-repo/AI integration/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python
requests
python-dotenv
httpx
google-generativeai
huggingface_hub
